+++ client/src/components/CommentAdd.css
@@ -1,0 +1,88 @@
+.modal-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.5); /* Voile gris semi-transparent */
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  z-index: 1000; /* Au-dessus des autres éléments */
+}
+
+.modal {
+  background: white;
+  padding: 20px;
+  border-radius: 12px;
+  box-shadow: 0 4px 10px rgba(0, 0, 0, 0.3);
+  width: 90%;
+  max-width: 500px;
+  position: relative;
+}
+
+.opinion-form {
+  display: flex;
+  flex-direction: column;
+  gap: 30px;
+  background: aliceblue;
+}
+
+.comment-section input {
+  width: 100%;
+  padding: 8px;
+  border: 1px solid #ccc;
+  border-radius: 5px;
+}
+
+.opinion-form {
+  background: ivory;
+  width: 800px;
+  height: 800px;
+  display: flex;
+  justify-content: left;
+  flex-direction: column;
+  gap: 20px;
+  border-radius: 8px;
+}
+
+.editor-section {
+  padding: 20px;
+}
+
+.modal-content .ql-container {
+  padding-bottom: 100px;
+}
+
+.modal-buttons,
+.exit-button {
+  display: flex;
+  justify-content: left;
+  gap: 20px;
+}
+
+.modal-buttons button {
+  padding: 10px 15px;
+  border: none;
+  cursor: pointer;
+  border-radius: 5px;
+  transition: 0.2s ease-in-out;
+}
+
+.modal-buttons button:first-child {
+  background-color: #4caf50;
+  color: white;
+}
+
+.modal-buttons button:last-child {
+  background-color: #f44336;
+  color: white;
+}
+
+.modal-buttons button:hover {
+  opacity: 0.8;
+
+  .modal-overlay {
+    background-color: aliceblue;
+  }
+}
